move the magic key dialog into a component
